Miscellaneous improvement in Form Widgets examples. Notably introduces an, as much as possible generalized (so far), dependent drop-downs example. I have introduced 2 couples of dependent drop-downs.
The 1st one, country/states, is a common one and does not need (ie it's better to not) to be generalized. It can be used as is. It needs some pieces which are all included, see comments and follow... 

The 2d, category/products, have been taken as example but is not a good candidate. What if you get much categ or/and products? Anyway it's a good OOTB example and I had it at hand, so I used it. I have tried to generalize (abstract the couple of dropdowns) and it's almost there. The only issue I crossed, is when you want to use a couple of dependent drop-downs in the same form. As it's not a common case, the setDependentDropdownValuesJs.ftl template-script can be used in most cases...

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/common@986370 13f79535-47bb-0310-9956-ffa450edef68