State dropdown only finds states and does not include county information (https://issues.apache.org/jira/browse/OFBIZ-4270) - I added a separte Entity which defines all countries which should show their states in the drop down list. It also checks states recursivly, to list all states (i.e.) for GB.
Orig Msg: If you are on this page https://demo-trunk.ofbiz.apache.org:8443/ecommerce/control/newcustomer select USA from the country drop down and it will display all the USA states, similarly it works just fine for Spain but if you try once again with either United Kingdom or France it will show "No States/Provinces exist".

Both UK and France have Counties saved in the datamodel but the form does not lookup in the correct location as they are not defined as State.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/common@1140103 13f79535-47bb-0310-9956-ffa450edef68