A modified patch from  Gareth Carter for "NotificationServices.java prepareNotification does not allow component:// urls" https://issues.apache.org/jira/browse/OFBIZ-5374
The prepareNotification method in NotificationServices.java does not allow urls that start with component:// because its using UtilURL rather than FlexibleLocation. Here is a patch that adds FlexibleLocation. I am not sure if its best to remove UtilURL or not - that can be up to you.

jleroux: since FlexibleLocation.resolveLocation() supersedes UtilURL.fromResource() I used it instead

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/common@1536905 13f79535-47bb-0310-9956-ffa450edef68