Workaround for "Lookups (layered mode) and dropdown menus overlap" (in BizznessTime Theme) https://issues.apache.org/jira/browse/OFBIZ-5519
A good example is in order creation page, open the customer lookup and then try to close it.

I think we should not fiddle with the theme, there is a faster and easier solution: center all lookups in the screen by default. It will not fix but hide the issue for now. Anyway the awful dropdown menu behaviour must be fixed and i will open another Jira for that.

You may ask why I'm concerned with a theme which is no longer bundled in future releases? It's because some custom projects still use it...


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/common@1563238 13f79535-47bb-0310-9956-ffa450edef68