A patch from Gil Portenseigne for "Field Lookup do not show description when widget-verbose is set to false" https://issues.apache.org/jira/browse/OFBIZ-6396
When widget verbose is set to false and using field lookup, the description of the selected object is not displayed (not using auto completion but the popup).
To reproduce : Set widget.verbose=false in framework/widget/config/widget.properties  
Clear OFBiz cache then go to : https://localhost:8443/partymgr/control/EditPartyRelationships?partyId=Company
In the partyRelationship creation form, in party Id From field, click on lookup, search, select AUTHOR_MADMAX, and so no description displayed. Using autocompletion it's fine.

This patch use a trick to no care about widgetVerbose comments. Sad to use this kind of trick, but if not, spaces might not be trimed. With that, the display is perfect. No side effect detected. 

jleroux: I still regret Hans past attitude about widgetVerbose which initially forced me to this hack 



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/common@1680873 13f79535-47bb-0310-9956-ffa450edef68